# 학생관리시스템 - 스크롤 tbody에만 적용 및 학생 삭제 요청

+++ src/main/java/com/example/studentmanager/student/Repository/StudentMapper.java
@@ -11,7 +11,7 @@ public interface StudentMapper {
     List<Student> findAll();
     boolean insert(Student stu);
     boolean update(Student stu);
-    boolean delete(int stuNum);
+    boolean delete(long stuNum);
 
     Student findOne(long num);
